+++ MTCH6102_ex.h
@@ -33,6 +33,15 @@
 // *****************************************************************************
 // Section: Types
 // *****************************************************************************
+typedef enum PACK__ MTCH6102_COMMAND {
+	MTCH_StoreToNV = 0x00,	//!< Write to non volatile storage
+	MTCH_RestoreDefaults,	//!< Restore controller to default configuration values
+	MTCH_Configure,			//!< Configure controller (after parameters have been changed)
+	MTCH_ManufacturingTest,	//!< Execute manufacturing test
+	MTCH_ForceBaseline,		//!< Force baseline (re-calibration) to occur
+} MTCH6102_COMMAND;
+
+
 typedef struct MTCH6102_raw {
 	uint8_t	sensor[16];
 } MTCH6102_raw_sense;
@@ -71,17 +80,85 @@ typedef struct MTCH6102_gesture {
 /****************************************/
 /*** High level methods and functions ***/
 /****************************************/
+__INLINE FctERR INLINE__ MTCH6102_Set_Mode(MTCH6102_MODE mode) {
+	if (mode > RawADC)	{ return ERROR_VALUE; }
+	return MTCH6102_Write(&mode, MTCH__MODE, 1); }
 
+FctERR MTCH6102_Command(MTCH6102_COMMAND cmd);
+
+__INLINE FctERR INLINE__ MTCH6102_Restore_Defaults(void) {
+	return MTCH6102_Command(MTCH_RestoreDefaults); }
+
+__INLINE FctERR INLINE__ MTCH6102_Store_To_Non_Volatile(void) {
+	return MTCH6102_Command(MTCH_StoreToNV); }
+
+__INLINE FctERR INLINE__ MTCH6102_Configuration_Request(void) {
+	return MTCH6102_Command(MTCH_Configure); }
+
+__INLINE FctERR INLINE__ MTCH6102_Manufacturing_Test(void) {
+	return MTCH6102_Command(MTCH_ManufacturingTest); }
+
+__INLINE FctERR INLINE__ MTCH6102_Force_Baseline(void) {
+	return MTCH6102_Command(MTCH_ForceBaseline); }
+
+
+/*** Touch Frame Control ***/
+__INLINE FctERR INLINE__ MTCH6102_Set_Active_Period(const uint16_t period) {
+	uint16_t per = perVal2perReg(period);
+	uint8_t DAT[2] = { LOBYTE(per), HIBYTE(per) };
+	return MTCH6102_Write(DAT, MTCH__ACTIVE_PERIOD_L, sizeof(DAT)); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Idle_Period(const uint16_t period) {
+	uint16_t per = perVal2perReg(period);
+	uint8_t DAT[2] = { LOBYTE(per), HIBYTE(per) };
+	return MTCH6102_Write(DAT, MTCH__IDLE_PERIOD_L, sizeof(DAT)); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Idle_Timeout(const uint8_t time) {
+	return MTCH6102_Write(&time, MTCH__IDLE_TIMEOUT, 1); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Debounce(const uint8_t touch, const uint8_t release) {
+	uint8_t DAT[2] = { release, touch };
+	return MTCH6102_Write(DAT, MTCH__DEBOUNCE_UP, sizeof(DAT)); }
+
+FctERR MTCH6102_Get_Active_Period(uint16_t * period);
+
+FctERR MTCH6102_Get_Idle_Period(uint16_t * period);
+
+
+/*** Acquisition and Touch Parameters ***/
+FctERR MTCH6102_Set_Filter(const MTCH6102_FILTER_TYPE type, const uint8_t strength, const bool base_filter);
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Scan_Count(const uint8_t count) {
+	return MTCH6102_Write(&count, MTCH__SCAN_COUNT, 1); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Touch_Threshold(const uint8_t xthr, const uint8_t ythr) {
+	uint8_t DAT[2] = { xthr, ythr };
+	return MTCH6102_Write(DAT, MTCH__TOUCH_THRESH_X, sizeof(DAT)); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Hysteresis(const uint8_t hyst) {
+	return MTCH6102_Write(&hyst, MTCH__HYSTERESIS, 1); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Large_Activation_Threshold(const uint16_t thr) {
+	uint8_t DAT[2] = { LOBYTE(thr), HIBYTE(thr) };
+	return MTCH6102_Write(DAT, MTCH__LARGE_ACTIVATION_THRESH_L, sizeof(DAT)); }
+
+
+/*** Baseline ***/
+__INLINE FctERR INLINE__ MTCH6102_Set_Base_Interval(const uint16_t interval) {
+	uint8_t DAT[2] = { LOBYTE(interval), HIBYTE(interval) };
+	return MTCH6102_Write(DAT, MTCH__BASE_INTERVAL_L, sizeof(DAT)); }
+
+__INLINE FctERR INLINE__ MTCH6102_Set_Base_Filter(const uint8_t pos, const uint8_t neg) {
+	uint8_t DAT[2] = { pos, neg };
+	return MTCH6102_Write(DAT, MTCH__BASE_POS_FILTER, sizeof(DAT)); }
+
+
+/*** Acquisition ***/
 __INLINE FctERR INLINE__ MTCH6102_Get_Raw(MTCH6102_raw_sense * raw) {
 	return MTCH6102_Read((uint8_t *) raw, MTCH__SENSOR_VALUE_RX0, sizeof(MTCH6102_raw_sense)); }
 
 __INLINE FctERR INLINE__ MTCH6102_Get_Gest(MTCH6102_raw_gest * gest) {
 	return MTCH6102_Read((uint8_t *) gest, MTCH__TOUCH_STATE, sizeof(MTCH6102_raw_gest)); }
-
-
-#define MTCH6102_raz()	{	uMTCH_REG__CMD cmd;		\
-							cmd.Bits.DEF = true;	\
-							MTCH6102_Write(&cmd, MTCH__CMD, 1); }
 
 
 /****************************************************************/