+++ Release Notes.md
@@ -25,6 +25,8 @@ SOFTWARE.
 * use of const qualifier in function parameters
 * init sequence returns with error value if something goes wrong during initialization
 * FM24C & MB85RC256V: added macros to create simple inlines to read/write values from addresses (to use in some header)
+* MTCH6102: added standalone function for commands & inlines for configuration, restore to defaults & tests commands
+* MTCH6102: added filtering type configuration function & inlines for configuration
 
 ## v0.6